miniplaces notebook works (VGG model, mediocre results), auxiliary scripts started but need improving, vqa notebook has a demo of the tool and idea of a netowkr.
miniplaces.ipynb downloads all of the data and necessary files, processes the data into single arrays for easy loading, and runs either VGG16 with whatever dense output layers you want or a basic CNN from scratch (the latter isn't really tested yet). The function to output the predictions on the test set seems to have an error, but I'm not sure what yet.

models.py contains the VGG model. This and the basic CNN model need to be combined somehow (e.g. have a superclass that the other CNNs inherit from) because they share a great deal of code.

run_models.py doesn't work yet; there's a decent scaffold, though some unneeded things in comments, but there's an unsolved error when you run it.

utils.py contains a couple of useful things like initialization for tensorflow and getting the next run number.

vqa.ipynb downloads the training data and does a demo of the VQA tools. There's no working model yet, but there is a bit of code that can be shaped into one.