run_models.py works now, models.py class has parameter for finetuning.
run_models.py was cleaned up a bit - mainly removing code that wasn't used. There's still a bit of old code around, though, that we might use later. Whatever the previous error was seems to be fixed. batch_size and log_key are now used properly. A new parameter, task, was added so that we can use the same script for miniplaces and vqa. The load_data function was updated to reflect the fact that there are multiple tasks with separate datasets (loading the vqa dataset isn't implemented yet, however).

models.py was updated to have a parameter for whether the CNN layers should be finetuned or not. The pretrained CNN module used is also a parameter now (VGG16 and Xception are supported; ResNet and Inception don't work because our images are too small).