BaseNN class created, PretrainedCNN and CNN updated to inherit from it.
models.py now has a class BaseNN which implements many of the methods that we'll want for multiple models (train, predict_proba, score, log, save, etc.). PretrainedCNN and CNN now inherit from this class and only redefine the build_graph function (as well as their additions to init to take care of model-specific parameters).

miniplaces.ipynb was updated to import and use these new models. It also now creates a TensorFlow config in an initialization cell, before loading the data, so that we can avoid out of memory errors from calling os.fork after the data has been loaded. This should be used in all models in the notebook instead of letting them create a new config internally.