Added check_graph method to BaseNN, accuracy methods to CNN, and miniplaces notebook now has a simple CNN with ~.2@1, .5@5.
check_graph ensures that the tensors that BaseNN expects for its methods exist in the graph (so that subclasses don't forget to implement them; they still have to remember to call check_graph, though).

The accuracy methods are the same as in PretrainedCNN but were previously missing from CNN (accuracy1, accuracy5).