Python3, added acc@1, acc@5 to tensorboard, early stop based on acc@5, tnrange now sets loss as description.
Starting with this commit, we're using python 3 instead of 2. The code won't be backwards compatible. Minor updates in models reflect this: there are now types for the parameters of CNN but not yet for other classes/functions.

acc@1 and acc@5 are computed on the dev set during training after each epoch through the training data.

Printing out the loss at the end of each epoch been removed, but the description of the batch range is updated to have the loss of the previous batch and the description of the epoch range shows the losses at the end of the last epoch.
We also use trange instead of range now if in_notebook is false. This means that you get a progress bar even when running the models from the command line.