Layers classes, Inception network, notebook updates.
The CNN class previously only allowed simple networks to be made - always using 3x3 convolutions with same padding with one max pooling layer (2x2) after each convolution layer. As convolution and other layers (such as pooling) may be mixed, specifying a list of parameters for each conv layer and lists of parameters for each other layer wasn't going to work well in general. Since we already needed to have a list that had parameters for each of the layers in order, it made sense to wrap that up in the structure of a class for each layer. layers.py contains these classes (ConvLayer, MaxPoolLayer, etc.). The format is fairly flexible and robust; not only can layers be stacked, they can also branch then merge again. Essentially, these are just nice wrappers around tensorflow layers/functions that make it easier to specify all of the parameters desired but then create the tensors themselves inside of the CNN class.

The CNN class was updated (in the commit before this one, actually, but I'm not going to go fix the message there now) to take a list of the layers as described above (and no longer accepts, e.g., dense_nodes or cnn_nodes) and create the appropriate graph structure from them. It also has type hints now that we're using Python 3.

Specifying a large network is not only unwieldy from the command line, it would also be too time consuming if you ever wanted to run it again. I've thus created run_inception.py which can be used instead of run_models.py to train the Inception v4 architecture. It's a good example of how nicely a graph can be specified with the Layer classes. I haven't finished running it yet, though, so there may still be errors during training.

The notebook has a few minor updates including adding a new model to the table of submitted models with slightly better performance than the first one.