Added DropoutLayer, improved layer saving.
Previously the layers that defined a kernel_initializer would throw errors when you tried to save them to an HDF5 file as part of the model's parameters. To fix this, I've moved instantiating the kernel_initializer to the apply function so that it isn't stored as an attribute of the Layer itself. In addition, all of the references to tensorflow functions/classes (e.g. the functions to call when applying a layer) have been moved to dictionaries. The Layers now have strings as parameters which are the keys for these dictionaries.