Added is_training tensor, using acc5 for early stopping, moved Inception to models.
Both batch_normalization and dropout behave differently depending on whether one is training the model or using it purely for inference. In order to support this behavior, the models now have an is_training tensor which defaults to False but is set true while running the train_op. This is also passed through to the Layers when they're applied so that they can use it appropriately internally.