Decayed learning rate, input image in [0, 1] instead of 0-255, use global pooling layer.
The learning rate is now exponentially decayed as training progresses (no parameter right now to switch this off). The input images are divided by 255 so their pixel values are in [0, 1]. The new global average pooling layer is used instead of an average pooling layer that only achieved the same result by knowing ahead of time the needed size. The optimizer has been changed to RMSProp, but this isn't necessarily permanent; ideally this should be a parameter to the model too. The inception network has been changed to have much less depth as part of testing why it's not learning.