Batch norm before non-linearities.
If a layer (e.g. ConvLayer or DenseLayer) has batch normalization after it, this is now computed after the affine transformation but before the non-linearity (previously it was after the non-linearity). This doesn't seem to have made a difference so far.
All of the Layers now have self.params, even if it is empty.