CNN2 (w/ dataset augmentation), logging to BSON, model names.
There's a second CNN class now that uses TFRecords files as input instead of being passed a numpy array. This class also supports dataset augmentation. It needs a bit more work to be able to easily generate predictions on a given TFRecords file.
The models now support (through a flag) logging either to an HDF5, as before, or to a BSON file (a specific one for each model) which can be read into MongoDB.
Instead of a `run_num`, models must now be given a `model_name`; this will be used similarly (as the name of the directory to create to save information about this model). This should make it easier to find different models in TensorBoard.