Rotation for image aug, _batch supports 1 iteration through a Dataset.
Note that the rotation used from tensorflow leaves black regions where the image doesn't cover; there's no option yet to fill these.
The Dataset in CNN2 is no longer repeated. If using a Dataset, _batch will run until the Dataset is exhausted and then return. This makes it easy to loop exactly once through a Dataset (which is useful for doing inference).