Updated notebook for abstract scenes.
The abstract scenes and their composition files are downloaded (this has also been moved into the script download.sh). The VQA demo has been partially but not fully changed to use the new images. There's png -> jpg conversion.
A demo of parsing object info from the scenes was added, but this info isn't stored yet.
Question preprocessing was updated and dataframes are made that have the questions, answers, multiple choices, and image paths.
Mappings are created for answer -> id and word -> id. Preparation of data for the models is done, but no models are run.
The test data isn't used at all yet.