Data loading and model building functions moved to vqa_utils.py.
load_data and the functions it depends on (process_objects, etc.) are all in this file. So are the functions for creating a predictions file and evaluating the multiple choice accuracy of the model. Also included are the model building functions such as soft_crop, memory_module, and get_layers. This makes it easier to run models both from notebooks and from scripts without needing to update all of these functions twice when changed.