Updated create_predictions_file to work w/o generator.
The generator is only needed if one is using image features in the model. Otherwise, the inputs can be passed in directly. There's a flag to determine if a generator should be made. Inputs are also now accepted as a parameter. This is useful if you want to load them with settings other than the default ones. If `inputs` is None, then the data will be loaded for the given split with the default settings, as before.